+++ utils/load.py
@@ -212,11 +212,16 @@ def model(model_architecture, model_class):
     return models[model_class][model_architecture]
 
 
-def optimizer(optimizer, momentum=0.0, dampening=0.0, nesterov=False):
+def optimizer(optimizer, momentum=0.0, dampening=0.0, nesterov=False, save_buffers=[]):
     optimizers = {
         "custom_sgd": (
             custom_sgd.SGD,
-            {"momentum": momentum, "dampening": dampening, "nesterov": nesterov},
+            {
+                "momentum": momentum,
+                "dampening": dampening,
+                "nesterov": nesterov,
+                "save_buffers": save_buffers,
+            },
         ),
         "sgd": (optim.SGD, {}),
         "momentum": (
